refactor(trace): migrate status component to v2
