+++ comp/trace/status/def/component.go
